fixed CMakeLists.txt, run_tests.sh; changed type help_t.occur to size_t*

+++ src/utils.h
@@ -27,7 +27,7 @@
  * appear:  Matrix storing allowed characters at each position
  */
 typedef struct help_t {
-  int occur[ALPHABET_SIZE];
+  size_t occur[ALPHABET_SIZE];
   bool option[ALPHABET_SIZE];
   char *forced;
   bool *appear;
